requirements.txt
Django
djangorestframework
psycopg2-binary
django-environ
factory_boy
django-filter